+++ R/transformations.R
@@ -381,6 +381,51 @@ combine_data  <- function(d = NULL, r = NULL, z = NULL, logOR = NULL, t = NULL, 
   }
 }
 
+.combine_data.bi <- function(x1, x2, n1, n2, study_names = NULL, study_ids = NULL){
+
+  BayesTools::check_int(x1[!is.na(x1)], "x1", check_length = FALSE,      lower = 0)
+  BayesTools::check_int(x2[!is.na(x2)], "x2", check_length = length(x1), lower = 0)
+  BayesTools::check_int(n1[!is.na(n1)], "n1", check_length = length(x1), lower = 0)
+  BayesTools::check_int(n2[!is.na(n2)], "n2", check_length = length(x1), lower = 0)
+  BayesTools::check_char(study_names, "study_names", allow_NULL = TRUE, check_length = length(x1))
+  BayesTools::check_char(study_names, "study_ids",   allow_NULL = TRUE, check_length = length(x1))
+
+  if(any(x1 > n1) | any(x2 > n2))
+    stop("Number of events cannot be larger than the number of observations")
+
+  # add study names if missing
+  if(is.null(study_names)){
+    study_names <- paste0("Study ", 1:length(x1))
+  }
+
+  # add study ids if missing
+  if(is.null(study_ids)){
+    study_ids <- rep(NA, length(x1))
+  }
+
+  # remove indicators from independent studies
+  study_ids[!study_ids %in% study_ids[duplicated(study_ids)]] <- NA
+  # assign factor levels
+  study_ids<- as.integer(as.factor(study_ids))
+
+  output <- data.frame(
+    x1          = x1,
+    x2          = x2,
+    n1          = n1,
+    n2          = n2,
+    study_names = study_names,
+    study_ids   = study_ids
+  )
+
+
+  attr(output, "effect_measure")   <- "freq"
+  attr(output, "outcome")          <- "freq"
+  attr(output, "all_independent")  <- all(is.na(output[,"study_ids"]))
+  attr(output, "weighted")         <- FALSE
+  class(output) <- c(class(output), "data.BiBMA")
+
+  return(output)
+}
 
 
 .transform_posterior       <- function(object, current_scale, output_scale){